+++ Kernel/TTY/VirtualConsole.cpp
@@ -131,7 +131,7 @@ inline bool is_valid_final_character(byte ch)
 unsigned parse_uint(const String& str, bool& ok)
 {
     unsigned value = 0;
-    for (size_t i = 0; i < str.length(); ++i) {
+    for (int i = 0; i < str.length(); ++i) {
         if (str[i] < '0' || str[i] > '9') {
             ok = false;
             return 0;
